+++ src/depack.js
@@ -1,6 +1,7 @@
-import loadScripts, { loadJSON } from './'
+import loadScripts, { loadJSON, loadStyle } from './'
 
 module.exports = {
   '_loadScripts': loadScripts,
   '_loadJSON': loadJSON,
+  '_loadStyle': loadStyle,
 }